Fix for badges being replaced multiple times
To reproduce
* start a stream
* let a few seconds play
* scrub back to the 0:00

The messages would be shuffled around
in the chat box and passed through the
observer a second time.

The badges would then have their "data" URL
passed into the function to get the badge ID,
and it would return undefined resulting in
a broken image icon being displayed.