Check author-box element exists before using it
The membership messages are initially added with
a <dom-if> element/template inside the inner column.
That is later replaced with the actual dom structure.
https://polymer-library.polymer-project.org/3.0/api/elements/dom-if

Can reproduce this error on stream OlaR3tnlg9g by
starting at 0 seconds, and before the initial data
has all been consumed jumping to the first highlight
at 10:16.

All the membership messages in the chat box will
initially appear as templates and so hit this error,
but they don't need the html replacing anyway as
they are patched during the fetch for those.